Rename Template resource to Profile
Profile is the domain-correct term (BIBFRAME/Dublin Core). Renames the
Template model/table/route/identity to Profile/profiles throughout:

- Template model -> Profile (table 'profiles', .../profiles/{uuid} uris)
- set_jsonld skips framing for Profile instances
- Migration 20260626 now creates the profiles table
