+++ src/bluecore_models/models/__init__.py
@@ -12,4 +12,4 @@ from bluecore_models.models.other_resource import (
     OtherResource as OtherResource,
     BibframeOtherResources as BibframeOtherResources,
 )
-from bluecore_models.models.template import Template as Template
+from bluecore_models.models.profile import Profile as Profile
